requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-widget-tweaks==1.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
